requirements.txt
django
gunicorn
djangorestframework
markdown
django-filter
psycopg2
django-oauth-toolkit
django-cors-middleware
